Rename "Rate Your Network" to "Network Rating"
it is easier to communicate the network ratings as a state than a verb, which is usually only understood in the moment.

Re-reading "Network Rating" conveys assurance rather than a repeated call to action.

Also made the label stick out more with 0.9 opacity - otherwise it is too hidden compared to other elements